Plugins: Add support for multiple plugins providing preload/postload code the same module
* This solves a long standing TODO and also makes the code slightly
  cleaner with more re-use.
